requirements.txt
pytest==7.4.4
pytest-html==4.1.1
pyyaml==5.3.1
playwright==1.41.1